+++ CHANGELOG.md
@@ -101,6 +101,9 @@ should not be broken.
 * Added `join()` template function. This is different from `separate()` in that
   it adds a separator between all arguments, even if empty.
 
+* `jj log` now supports a `--count` flag to print the number of commits instead
+  of displaying them.
+
 ### Fixed bugs
 
 * `jj fix` now prints a warning if a tool failed to run on a file.